requirements.txt
flask
openai
python-dotenv
flask-socketio
